+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Introduce Surveys BB [#2703](https://github.com/rokwire/illinois-app/issues/2703)
 ### Fixed
 - Style definition for wellness to-do item card [#2702](https://github.com/rokwire/illinois-app/issues/2702)